feat: Move testDebug to test
Move the logging-api module's testDebug directory to test

Resolves: 8381
