+++ modules/logging-api/src/test/java/uk/gov/logging/api/analytics/extensions/StringExtensionsTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 
-internal class StringExtensionsKtTest {
+class StringExtensionsTest {
 
     @ParameterizedTest(name = "{index}: {0}")
     @MethodSource("setupMd5Output")
