+++ modules/logging-api/src/test/java/uk/gov/logging/api/analytics/parameters/v2/ButtonParametersTest.kt
@@ -7,9 +7,9 @@ import uk.gov.logging.api.analytics.logging.TEXT
 import uk.gov.logging.api.analytics.logging.TYPE
 import uk.gov.logging.api.analytics.parameters.ParametersTestData
 import uk.gov.logging.api.analytics.parameters.data.TaxonomyLevel2
-import uk.gov.logging.api.analytics.parameters.data.TypeIcon
+import uk.gov.logging.api.analytics.parameters.data.TypeSubmitForm
 
-internal class IconParametersTest {
+class ButtonParametersTest {
 
     private val exampleCallToActionText = "button text"
 
@@ -17,9 +17,9 @@ internal class IconParametersTest {
     fun `text is truncated to be 100 characters or less`() {
         assertEquals(
             ParametersTestData.overOneHundredString.lowercase().take(HUNDRED_CHAR_LIMIT),
-            IconParameters(
+            ButtonParameters(
                 text = ParametersTestData.overOneHundredString,
-                type = TypeIcon.ICON,
+                type = TypeSubmitForm.SUBMIT_FORM,
                 overrides = RequiredParameters(
                     taxonomyLevel2 = TaxonomyLevel2.GOVUK
                 )
@@ -31,12 +31,12 @@ internal class IconParametersTest {
     fun `Match output map`() {
         val expectedMap = mutableMapOf<String, Any?>(
             TEXT to exampleCallToActionText.lowercase(),
-            TYPE to TypeIcon.ICON.value
+            TYPE to "submit form"
         )
 
-        val mapper = IconParameters(
+        val mapper = ButtonParameters(
             text = exampleCallToActionText,
-            type = TypeIcon.ICON,
+            type = TypeSubmitForm.SUBMIT_FORM,
             overrides = RequiredParameters(
                 taxonomyLevel2 = TaxonomyLevel2.GOVUK
             )
